Refactor: asset 관련 유틸 함수 경로 수정 및 정책 관련 코드 삭제 (#22)
