+++ utils/asset/get_iam.py
@@ -1,4 +1,3 @@
-from datetime import datetime, timedelta, timezone
 import boto3
 from dotenv import load_dotenv
 import os
